% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedClustSuite.R
\name{DensityPeakClustered}
\alias{DensityPeakClustered}
\alias{DensityPeakClustered.ClusteringObject}
\title{Check whether a DensityPeake ClusteringObject object have been DensityPeakClustered}
\usage{
DensityPeakClustered(x)

\method{DensityPeakClustered}{ClusteringObject}(x)
}
\arguments{
\item{x}{A ClusteringObject object}
}
\value{
\code{TRUE} if \code{\link[=findDensityPeakClusters]{findDensityPeakClusters()}} have been performed, otherwise
\code{FALSE}
}
\description{
This function checks whether \code{\link[=findDensityPeakClusters]{findDensityPeakClusters()}} has been performed on
the given object and returns a boolean depending on the outcome
}
