% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedClustSuite.R
\name{findDensityPeakClusters}
\alias{findDensityPeakClusters}
\alias{findDensityPeakClusters.ClusteringObject}
\title{Detect clusters in a  DensityPeakity Cluster obejct}
\usage{
findDensityPeakClusters(x, ...)

\method{findDensityPeakClusters}{ClusteringObject}(
  x,
  rho,
  delta,
  plot = FALSE,
  peaks = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{rho}{The threshold for local density when detecting cluster peaks}

\item{delta}{The threshold for minimum distance to higher density when detecting cluster peaks}

\item{plot}{Logical. Should a decision plot be shown after cluster detection}

\item{peaks}{A numeric vector indicates the index of density peaks used for clustering. This vector should be retrieved from the decision plot with caution. No checking involved.}

\item{verbose}{Logical. Should the running details be reported}
}
\description{
Detect clusters in a  DensityPeakity Cluster obejct
}
\examples{
irisDist <- dist(iris[,1:4])
irisClust <- ClustObj(irisDist, gaussian=TRUE)
plot(irisClust) # Inspect clustering attributes to define thresholds

irisClust <- findDensityPeakClusters(irisClust, rho=2, delta=2)
plotMDS(irisClust)
split(iris[,5], irisClust$clusters)
}
