% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedClustSuite.R
\name{plotTSNE}
\alias{plotTSNE}
\title{Plot observations using t-distributed neighbor embedding and colour by cluster}
\usage{
plotTSNE(x, ...)
}
\description{
Plot observations using t-distributed neighbor embedding and colour by cluster
}
\examples{
irisDist <- dist(iris[,1:4])
irisClust <- ClustObj(irisDist, gaussian=TRUE)
plot(irisClust) # Inspect clustering attributes to define thresholds

irisClust <- findDensityPeakClusters(irisClust, rho=2, delta=2)
plotTSNE(irisClust)
split(iris[,5], irisClust$clusters)
}
