% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedClustSuite.R
\docType{package}
\name{weightedClustSuite-package}
\alias{weightedClustSuite}
\alias{weightedClustSuite-package}
\title{weightedClustSuite: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\examples{
irisDist <- dist(iris[,1:4])
irisClust <- ClustObj(irisDist, gaussian=TRUE)
plot(irisClust) # Inspect clustering attributes to define thresholds

irisClust <- findDensityPeakClusters(irisClust, rho=2, delta=2)
plotMDS(irisClust)
split(iris[,5], irisClust$clusters)

}
\author{
\strong{Maintainer}: Dhanuj Gandikota \email{dhanujg@umich.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
